allow overflow scrolling on my-programs view * new dialog for creating a program
