+++ src/view/components/Common/List.css
@@ -1,4 +1,5 @@
 .list {
+  overflow-y: auto;
   list-style: none;
   padding: 0;
   margin: 0;
